feat(ui): made separate css file for 404 page
