+++ frontend/src/pages/404/NotFound.jsx
@@ -1,3 +1,4 @@
+import './NotFound.css';
 import { useNavigate } from 'react-router-dom';
 
 export default function NotFound() {
